betvision_backend/requirements.txt
# BetVision Pro Backend Requirements

# Django Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL (optional)

# Image Processing
Pillow>=10.0.0

# Web Scraping & Monitoring
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Flask (for existing monitoring services)
Flask>=2.3.0
Flask-CORS>=4.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development
python-dotenv>=1.0.0